Autonomous algorithm works, and it's not infinite!
So we have a counter for how many moves the robot makes. If it exceeds a
threshold and hasn't hit a low enough RSSI value, it'll reset the RSSI
threshold to it's current local max with the diff factored in. Then
since we've hit that local max before, guaranteed to find it again.
WHEE. Also Android app knows when in autonomous mode and when in manual.
Currently toast signals end of autonomous mode.